prgtest 52_23
Hora: 2024-05-13 16:35:07

+++ 52_23_elimina_total/Client.java
@@ -69,8 +69,8 @@ public class Client {
         return resultat;
     }
 
-    private int importTotal() {
-        int total = 0;
+    private double importTotal() {
+        double total = 0;
         for (Lloguer lloguer : lloguers){
             total += lloguer.quantitat() * 30;
         }
